refactor(utils): extract path functions from sidebar
